requirements.txt
requests~=2.31.0
python-dotenv~=1.0.0
oauthlib~=3.2.2